+++ app/src/main/java/com/blastbet/nanodegree/popularmovies/tmdb/MovieApi.java
@@ -13,7 +13,7 @@ public interface MovieApi {
     public Call<MovieList> getMovieList(@Path("sortorder") String sortOrder, @Query("api_key") String apiKey);
 
     @GET("movie/{id}")
-    public Call<Movie> getMovieDetails(@Path("id") int movieId, @Query("api_key") String apiKey);
+    public Call<Movie> getMovieDetails(@Path("id") long movieId, @Query("api_key") String apiKey);
 
     @GET("movie/{id}/videos")
     public Call<MovieTrailerList> getMovieTrailers(@Path("id") int movieId, @Query("api_key") String apiKey);
